// src/schemas/ApplicationStatus.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Application status flags */
export type ApplicationStatus = {
  /** Standard ROS header */
  header: Header;

  /** Speed limit 4kmh status */
  app_status_speed_limit4kmh: boolean;

  /** CAN debug messages enabled */
  app_status_send_c_a_n_dbg_messages: boolean;

  /** Power assisted braking enabled */
  app_status_pwr_assisted_braking: boolean;
};