// src/schemas/RawSignalThrottle.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Raw throttle signal */
export type RawSignalThrottle = {
  /** Standard ROS header */
  header: Header;

  uint32_signal_throttle_b: number;
  uint32_signal_throttle_a: number;
};