// src/schemas/RemoteDriveRequest.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Remote drive request */
export type RemoteDriveRequest = {
  /** Standard ROS header */
  header: Header;

  /** Remote steering angle request */
  remote_steering_angle_req: number;

  /** Remote velocity request (m/s) */
  remote_velocity_req: number;
};