// src/schemas/ScaledSignals.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Scaled control signals */
export type ScaledSignals = {
  /** Standard ROS header */
  header: Header;

  throttle_signal: number;
  steering_motor_speed_cmd: number;
  curtis_speed_cmd: number;
  steering_torque_signal: number;
  steering_velocity_signal: number;
  brake_signal: number;
};