// src/schemas/SteeringAndSpeed.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Steering and speed signals */
export type SteeringAndSpeed = {
  /** Standard ROS header */
  header: Header;

  /** Vehicle velocity requested (m/s) */
  vehicle_velocity_requested: number;

  /** Steering position requested (percent) */
  steering_position_requested: number;

  /** Vehicle velocity measured (m/s) */
  vehicle_velocity_measured: number;

  /** Steering position measured (percent) */
  steering_position_measured: number;
};